revamp create-project flow
