+++ components/project/post/InterestsPage.tsx
@@ -2,8 +2,9 @@ import React, { useState } from 'react';
 import { Box, Typography } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 
+import HelpButton from 'components/HelpButton';
 import ArrayForm from 'components/ArrayForm';
-import { FONT, COLORS } from 'public/static/styles/constants';
+import { COLORS } from 'public/static/styles/constants';
 
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -13,12 +14,10 @@ const useStyles = makeStyles((theme) => ({
     justifyContent: 'space-between',
     padding: theme.spacing(4),
   },
-  text: {
-    fontSize: FONT.GUIDE,
-    color: theme.palette.text.secondary,
-  },
   title: {
     paddingLeft: theme.spacing(1),
+    marginRight: theme.spacing(5),
+    fontWeight: 'bold',
     color: theme.palette.text.primary,
   },
 }));
@@ -30,16 +29,16 @@ export default ({ project, options }): JSX.Element => {
 
   return (
     <Box className={classes.container}>
-      <Box paddingX="1rem" marginTop="1rem" marginBottom="3rem">
-        <Typography align="center" className={classes.text}>
-          {`My project categories belonged to ‘healthcare’, and ‘IT’ as we built a web that provides health information.`}
-        </Typography>
-      </Box>
       <Box minHeight="80%">
-        <Typography className={classes.title}>
-          <span style={{ fontWeight: 'bold' }}>{`Areas of Interest`}</span>
-          <span style={{ color: COLORS.GRAY_BB }}>{` (max 5)`}</span>
-        </Typography>
+        <Box display="flex">
+          <Typography className={classes.title}>
+            <span style={{ marginRight: '.5rem' }}>{`Areas of Interest`}</span>
+            <span style={{ color: COLORS.GRAY_BB, fontWeight: 'normal' }}>{` (max 5)`}</span>
+          </Typography>
+          <HelpButton
+            text={`This helps users understand the field of studies where your project could fall under. Good examples are 'healthcare', 'AI', 'politics', and 'international studies'.`}
+          />
+        </Box>
         <ArrayForm label="Interests" items={interests} setItems={setInterests} limit={5} options={options} />
       </Box>
     </Box>
