+++ pages/project/[pid].tsx
@@ -1,25 +1,17 @@
 import React, { useState } from 'react';
 import { NextPage } from 'next';
 import Router from 'next/router';
-import { Container, Box, Grid, IconButton, Modal, Fade, Typography } from '@material-ui/core';
+import { Container, Box, Grid, IconButton } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 
 import { BE_ADDR, FE_ADDR, callApi, redirectPage, vh } from 'utils';
 import MainButton from 'components/MainButton';
 import ProjectContent from 'components/project/ProjectContent';
+import useModal from 'components/InfoModal';
 import { Project, Contract } from 'types';
-import { FONT, COLORS } from 'public/static/styles/constants';
+import { COLORS } from 'public/static/styles/constants';
 
 const useStyles = makeStyles((theme) => ({
-  content: {
-    display: 'flex',
-    flexDirection: 'column',
-    paddingBottom: theme.spacing(20),
-    '& > *': {
-      paddingLeft: theme.spacing(5),
-      paddingRight: theme.spacing(5),
-    },
-  },
   heading: {
     padding: theme.spacing(4),
     paddingTop: theme.spacing(2),
@@ -40,21 +32,6 @@ const useStyles = makeStyles((theme) => ({
     backgroundColor: 'white',
     borderTop: `1px solid ${COLORS.GRAY_DA}`,
   },
-  modal: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  modalBox: {
-    backgroundColor: 'white',
-    boxShadow: '0px 12px 48px rgba(0, 0, 0, 0.18)',
-    borderRadius: theme.spacing(1),
-    color: COLORS.GRAY_75,
-    fontSize: FONT.GUIDE,
-    maxWidth: '300px',
-    padding: '.5rem',
-    paddingBottom: '2rem',
-  },
 }));
 
 type PageProps = {
@@ -69,7 +46,7 @@ type PageProps = {
 const ProjectPage: NextPage<PageProps> = ({ project, projectId, saved, isConnected }) => {
   const classes = useStyles();
   const [saveStatus, setSaveStatus] = useState(saved);
-  const [showModal, setShowModal] = useState(false);
+  const [InfoModal, setShowModal] = useModal();
 
   const handleJoinRequest = async (event): Promise<void> => {
     event.preventDefault();
@@ -110,7 +87,7 @@ const ProjectPage: NextPage<PageProps> = ({ project, projectId, saved, isConnect
         </Grid>
       </Container>
 
-      <Container component="main" maxWidth="xs" disableGutters className={classes.content}>
+      <Container component="main" maxWidth="xs" disableGutters>
         <ProjectContent project={project} />
       </Container>
 
@@ -122,29 +99,9 @@ const ProjectPage: NextPage<PageProps> = ({ project, projectId, saved, isConnect
               onClick={handleJoinRequest}
               disabled={isConnected}
             />
-
-            <Modal
-              className={classes.modal}
-              open={showModal}
-              onClose={(): void => setShowModal(false)}
-              closeAfterTransition
-              BackdropProps={{ invisible: true }}
-            >
-              <Fade in={showModal}>
-                <div className={classes.modalBox}>
-                  <Box display="flex" justifyContent="flex-end">
-                    <IconButton style={{ padding: '0' }} onClick={(): void => setShowModal(false)}>
-                      <img src="/static/assets/modal_exit.svg" alt="modal-exit" />
-                    </IconButton>
-                  </Box>
-                  <Box padding=".5rem">
-                    <Typography>
-                      {`Your email has been shared with the project manager. He/she will contact you if they think you are a good fit!`}
-                    </Typography>
-                  </Box>
-                </div>
-              </Fade>
-            </Modal>
+            <InfoModal
+              text={`Your email has been shared with the project manager. He/she will contact you if they think you are a good fit!`}
+            />
           </Box>
         </Container>
       </Box>
